% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_functions.R
\name{read_and_join}
\alias{read_and_join}
\title{read in a list of files with the same column format}
\usage{
read_and_join(file_list, header = TRUE)
}
\arguments{
\item{file_list}{vector file locations.}

\item{header}{Do the files have a header?}
}
\value{
data.frame of all files concatenated using rbind()
}
\description{
read in a list of files with the same column format
}
\author{
Beth Signal
}
