% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_functions.R
\name{read_fasta2df}
\alias{read_fasta2df}
\title{read in a fasta file as a data.frame}
\usage{
read_fasta2df(file)
}
\arguments{
\item{file}{fasta file location.}
}
\value{
data.frame fasta sequences. One row for each sequence
}
\description{
read in a fasta file as a data.frame
}
\author{
Beth Signal
}
