% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{remove_quotes}
\alias{remove_quotes}
\title{remove any stray quotes in strings}
\usage{
remove_quotes(string)
}
\arguments{
\item{string}{string}
}
\value{
string without quotes
}
\description{
remove any stray quotes in strings
}
\examples{
remove_quotes("abcd'efgh")
remove_quotes('abcd"efgh')
}
\author{
Beth Signal
}
