% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_functions.R
\name{strv_split2}
\alias{strv_split2}
\title{apply str_split to string vectors to extract text between two delimters}
\usage{
strv_split2(string_vector, split_by_1, split_by_2, position_1 = NULL,
  position_2 = NULL)
}
\arguments{
\item{string_vector}{vector of strings}

\item{split_by_1}{first character delimiter to split strings by (at the start of the desired string to be captured)}

\item{split_by_2}{second character delimiter to split strings by (at the end of the desired string to be captured)}

\item{position_1}{which position of the first split to return}

\item{position_2}{which position of the second split to return}
}
\value{
vector of strings split by the split_by character at the position
}
\description{
apply str_split to string vectors to extract text between two delimters
}
\examples{
strv_split2(c("ABC_DEF_HIJ_KLM", "ABC_QRS_HIJ_WXY"), "C_" , "_H")
}
\author{
Beth Signal
}
